lints: fix moved lint_ian_bare_wildcard.go source/category/package
