+++ lints/community/lint_ian_bare_wildcard.go
@@ -1,4 +1,4 @@
-package rfc
+package community
 
 /*
  * ZLint Copyright 2020 Regents of the University of Michigan
@@ -45,8 +45,8 @@ func init() {
 	lint.RegisterLint(&lint.Lint{
 		Name:          "e_ian_bare_wildcard",
 		Description:   "A wildcard MUST be accompanied by other data to its right (Only checks IANDNSNames)",
-		Citation:      "RFC5280",
-		Source:        lint.RFC5280,
+		Citation:      "awslabs certlint",
+		Source:        lint.AWSLabs,
 		EffectiveDate: util.ZeroDate,
 		Lint:          &brIANBareWildcard{},
 	})
